requirements.txt
# IoT API Integration Dependencies
requests==2.31.0
pyodbc==5.0.1
python-dotenv==1.0.0
pandas==2.1.4
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
schedule==1.2.0
